+++ opentelemetry-http/src/lib.rs
@@ -83,6 +83,7 @@ mod reqwest {
         }
     }
 
+    #[cfg(not(target_arch = "wasm32"))]
     #[async_trait]
     impl HttpClient for reqwest::blocking::Client {
         async fn send(&self, request: Request<Vec<u8>>) -> Result<Response<Bytes>, HttpError> {